# docs/swagger.yaml
basePath: /
definitions:
  app.loginReq:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  app.registerReq:
    properties:
      login:
        type: string
      name:
        type: string
      pass:
        type: string
    type: object
  app.registerResp:
    properties:
      ok:
        type: boolean
    type: object
  ds.Reagent:
    properties:
      description:
        type: string
      formula:
        type: string
      id:
        type: integer
      img:
        type: string
      molar_mass:
        type: number
      name:
        type: string
      video:
        type: string
    type: object
host: 127.0.0.1:8080
info:
  contact:
    email: bitop@spatecon.ru
    name: API Support
    url: https://vk.com/bmstu_schedule
  description: Bmstu Open IT Platform
  license:
    name: AS IS (NO WARRANTY)
  title: BITOP
  version: "1.0"
paths:
  /api/reagents:
    get:
      description: Публичный метод чтения данных
      parameters:
      - description: Поиск
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ds.Reagent'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Список реагентов
      tags:
      - reagents
  /auth/login:
    post:
      consumes:
      - application/json
      description: Создаёт серверную сессию в Redis и устанавливает cookie session_id
      parameters:
      - description: Данные входа
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/app.loginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Вход пользователя
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Создаёт нового пользователя-исследователя
      parameters:
      - description: Данные регистрации
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/app.registerReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.registerResp'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Регистрация пользователя
      tags:
      - auth
schemes:
- http
securityDefinitions:
  SessionCookieAuth:
    in: cookie
    name: session_id
    type: apiKey
swagger: "2.0"